backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic-settings==2.7.1
httpx==0.28.1
python-multipart==0.0.20
alembic==1.14.1
pytest==8.3.4
pytest-asyncio==0.25.0
